KTB-17 Реализовать функционал отображения статистики в боте. Привёл в порядок проект
