--- src/main/kotlin/LEB_02/LEB_02_01.kt
@@ -1,8 +1,0 @@
-package LEB_02
-
-import java.io.File
-
-fun main() {
-    val wordsFile: File = File("words.txt")
-    wordsFile.forEachLine { println(it) }
-}
