+++ src/main/kotlin/ru/androidsprint/englishtrainer/treaner/LearnWordsTrainer.kt
@@ -1,4 +1,4 @@
-package LEB_04
+package ru.androidsprint.englishtrainer.treaner
 
 import java.io.File
 import java.lang.IllegalStateException
